Move dummy scheduler into test suite, AsyncDataloader requires a scheduler
